use size_t for iteration in motor_hardware

+++ src/motor_hardware.cc
@@ -45,7 +45,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 MotorHardware::MotorHardware(ros::NodeHandle nh){
 	ros::V_string joint_names = boost::assign::list_of("left_wheel_joint")("right_wheel_joint");
 
-	for (unsigned int i = 0; i < joint_names.size(); i++) {
+	for (size_t i = 0; i < joint_names.size(); i++) {
 		hardware_interface::JointStateHandle joint_state_handle(joint_names[i],
 		    &joints_[i].position, &joints_[i].velocity, &joints_[i].effort);
 		joint_state_interface_.registerHandle(joint_state_handle);
